requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
python-dotenv
pydantic
pydantic-settings
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
